requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
anthropic>=0.40.0
PyPDF2>=3.0.0
python-docx>=1.0.0
Pillow>=10.0.0
supabase>=2.0.0
